Themes: Update ThemesContract with install state [2/3]
This provider will be responsible for keeping track of the installed
state of themes.  This includes two intermedite states, INSTALLING
and UPDATING.  When a theme is installed from PackageManager the
provider will receive this broadcast and add the theme to its DB
and set the state to INSTALLING if the ThemeService is processing
the theme, or INSTALLED if not and broadcast the ACTION_THEME_INSTALLED
action.  When a theme is updated the provider will receive another
broadcast indicating that the current version of the theme is being
replaced.  If the ThemeService needs to do some processing, we'll
update the state to UPDATING.

Once the theme service is done processing a theme, the provider
will receive a broadcast from the ThemeService.  At this point, the
themes provider will update the state to INSTALLED, if it is not
already set to that state.  The provider will broadcast the
ACTION_THEME_INSTALLED action if the previous state was INSTALLING
and the ACTION_THEME_UPDATED action if the previous state was
UPDATING.

Change-Id: Ie37f85463e472d96b8393801007537db1e3eefc6