Avoid OOME when generating many wallpaper previews
A pack containing a large collection of wallpapers could cause the
preview generation to fail with an OOME since we were storing all
the previews in memory before writing them to disk and adding their
entries into the database.

Wallpaper previews are now stored on disk within the wallpaper
preview generator and the path is returned in the WallpaperItem
object instead of the bitmaps.

Change-Id: I1dca215bc2fab9266c0baa4bce2a39e2ff1238a6